pyproject.toml
[tool.poetry]
name = "fatcat-worker"
version = "0.1.0"
description = ""
authors = ["Yotamefr <yotamefr@gmail.com>"]
readme = "README.md"
packages = [{include = "fatcat_worker"}]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.1"
gitpython = "^3.1.32"
python-dotenv = "^1.0.0"
argparse = "^1.4.0"
fatcat-utils = {git = "https://github.com/Yotamefr/fatcat-utils", rev = "dev"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pylint = "^2.17.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
